feat(plasma-tokens-utils): Added `WebColors` list
